Модуляция: разделила на модули, переписала функции генерации
